Add option "datesHighlighted"
Added an additional option to allow the user to add custom classes to any dates with the "datesHighlighted" option, which takes an object in the form of { date : class(es) }.  Dates are parsed in Datepicker standard method.  Classes can be passed in as raw strings, arrays, and comma-separated lists.  Usage:
$('#sandbox-container div').datepicker({
    datesHighlighted: { '07/06/2012' : 'birthday', '03/21/2004' : 'holiday' }
});